+++ src/pbfbench/topics/assembly/description.py
@@ -2,4 +2,4 @@
 
 import pbfbench.abc.topic.description as abc_topic_desc
 
-DESCRIPTION = abc_topic_desc.Description("ASSEMBLY", "asm")
+DESCRIPTION = abc_topic_desc.Description("ASSEMBLY", "assembly")