+++ CHANGELOG.md
@@ -24,6 +24,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 <!-- next-header -->
+## [Unreleased] - yyyy-mm-dd
+
+### Changed
+
+* Make `work_exp_fs_manager` an object attribute of `abc_tool_res_items.Result`
+
 ## [0.2.0] - 2025-05-09
 
 ### Changed